# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "API Donadores y Entidades - 2026 - DDSI - TPA",
    "description": "API del componente de Donadores y Entidades - Lola Simone",
    "version": "v1"
  },
  "servers": [
    {
      "url": "https://entrega-2-lolasimone.onrender.com",
      "description": "Servidor en Render"
    }
  ],
  "paths": {
    "/donadores": {
      "post": {
        "summary": "Agregar un donador",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DonadorDTO" } } }
        },
        "responses": { "200": { "description": "Donador creado correctamente" } }
      },
      "get": {
        "summary": "Obtener todos los donadores",
        "responses": { "200": { "description": "Lista de donadores" } }
      }
    },
    "/donadores/{id}": {
      "get": {
        "summary": "Buscar donador por ID",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Donador encontrado" }, "404": { "description": "Donador no encontrado" } }
      }
    },
    "/donadores/{id}/estado": {
      "patch": {
        "summary": "Modificar estado del donador",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "properties": { "estado": { "type": "string", "enum": ["VERIFICADO", "SOSPECHOSO", "BANEADO"] } } } } }
        },
        "responses": { "200": { "description": "Estado actualizado" } }
      }
    },
    "/donadores/{id}/categoria": {
      "patch": {
        "summary": "Modificar categoría del donador",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "properties": { "categoria": { "type": "string" } } } } }
        },
        "responses": { "200": { "description": "Categoría actualizada" } }
      }
    },
    "/donadores/{id}/puede-donar": {
      "get": {
        "summary": "Verificar si un donador puede donar",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Resultado de validación" } }
      }
    },
    "/donadores/{id}/estadisticas": {
      "get": {
        "summary": "Obtener estadísticas del donador",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Estadísticas obtenidas" } }
      }
    },
    "/donadores/{id}/quejas": {
      "post": {
        "summary": "Agregar una queja a un donador",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/QuejaDTO" } } }
        },
        "responses": { "201": { "description": "Queja creada correctamente" } }
      },
      "get": {
        "summary": "Obtener quejas de un donador",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Lista de quejas" } }
      }
    },
    "/entidades": {
      "post": {
        "summary": "Agregar entidad benéfica",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EntidadDTO" } } }
        },
        "responses": { "201": { "description": "Entidad creada" } }
      },
      "get": {
        "summary": "Obtener todas las entidades",
        "responses": { "200": { "description": "Lista de entidades" } }
      }
    },
    "/entidades/{id}": {
      "get": {
        "summary": "Buscar entidad por ID",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Entidad encontrada" } }
      }
    },
    "/necesidades": {
      "post": {
        "summary": "Registrar necesidad material",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/NecesidadDTO" } } }
        },
        "responses": { "201": { "description": "Necesidad creada" } }
      },
      "get": {
        "summary": "Obtener necesidades por producto",
        "parameters": [{ "name": "productoID", "in": "query", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Lista de necesidades" } }
      }
    },
    "/necesidades/{necesidadID}/satisfaccion": {
      "post": {
        "summary": "Satisfacer una necesidad",
        "parameters": [
          { "name": "necesidadID", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "cantidad", "in": "query", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "Necesidad actualizada" } }
      }
    }
  },
  "components": {
    "schemas": {
      "DonadorDTO": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "nombre": { "type": "string" },
          "apellido": { "type": "string" },
          "email": { "type": "string" },
          "estado": { "type": "string", "enum": ["VERIFICADO", "SOSPECHOSO", "BANEADO"] },
          "categoria": { "type": "string" }
        }
      },
      "EntidadDTO": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "razonSocial": { "type": "string" },
          "correo": { "type": "string" }
        }
      },
      "NecesidadDTO": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "entidadID": { "type": "string" },
          "descripcion": { "type": "string" },
          "cantidadObjetivo": { "type": "integer" },
          "productoID": { "type": "string" }
        }
      },
      "QuejaDTO": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "donadorID": { "type": "string" },
          "descripcion": { "type": "string" },
          "fecha": { "type": "string", "format": "date" }
        }
      }
    }
  }
}